src/requirements.txt
altair==5.3.0
asgiref==3.8.1
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
Django==5.0.6
easydict==1.13
filelock==3.13.1
fonttools==4.51.0
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.23.1
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.1.0
protobuf==5.27.2
psutil==6.0.0
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==2.4.0
six==1.16.0
smmap==5.0.1
sqlparse==0.5.0
streamlit==1.36.0
sympy==1.12
tbb==2021.12.0
tenacity==8.4.2
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.4.1
tqdm==4.66.4
transformers==4.41.1
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.0
watchdog==4.0.1
